@@ -3,7 +3,7 @@ using MyCourse.Models.Enums;
 
 namespace MyCourse.Models.ValueTypes
 {
-    public class Money
+    public record Money
     {
         public Money() : this(Currency.EUR, 0.00m)
         {
@@ -20,7 +20,7 @@ namespace MyCourse.Models.ValueTypes
             {
                 return amount;
             }
-            set
+            init
             {
                 if (value < 0) {
                     throw new InvalidOperationException("The amount cannot be negative");
@@ -30,22 +30,9 @@ namespace MyCourse.Models.ValueTypes
         }
         public Currency Currency
         {
-            get; set;
+            get; init;
         }
 
-        public override bool Equals(object obj)
-        {
-            var money = obj as Money;
-            return money != null &&
-                   Amount == money.Amount &&
-                   Currency == money.Currency;
-        }
-
-        public override int GetHashCode()
-        {
-            return HashCode.Combine(Amount, Currency);
-        }
-        
         public override string ToString()
         {
             return $"{Currency} {Amount:#.00}";
